pyproject.toml
[project]
name = "resume-multiagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "langchain>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-tavily>=0.2.13",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
]
